Custom hook added to HomePage in order to let the component more clean
